+++ src/StarPrinterInformationBluetoothLE.d.ts
@@ -1,6 +1,8 @@
 export declare class StarPrinterInformationBluetoothLE {
     _address: string | undefined;
     _deviceName: string | undefined;
+    _rssi: number | undefined;
     get address(): string | undefined;
     get deviceName(): string | undefined;
+    get rssi(): number | undefined;
 }